Seperate timer and pattern
