+++ flux_led/__init__.py
@@ -1,20 +1,20 @@
 """Init file for Flux LED"""
 from .fluxled import (
     DeviceType,
-    PresetPattern,
-    LedTimer,
     WifiLedBulb,
-    utils,
 )
 
 
 from .scanner import bulbscanner
+from .pattern import presetpattern
+from .timer import ledtimer
+from .utils import utils
 
 __all__ = [
     "DeviceType",
-    "PresetPattern",
-    "LedTimer",
+    "presetpattern",
+    "ledtimer",
     "WifiLedBulb",
-    "BulbScanner",
+    "bulbscanner",
     "utils",
 ]
