+++ flux_led/pattern.py
@@ -1,0 +1,36 @@
+
+
+class presetpattern:
+    seven_color_cross_fade = 0x25
+    red_gradual_change = 0x26
+    green_gradual_change = 0x27
+    blue_gradual_change = 0x28
+    yellow_gradual_change = 0x29
+    cyan_gradual_change = 0x2A
+    purple_gradual_change = 0x2B
+    white_gradual_change = 0x2C
+    red_green_cross_fade = 0x2D
+    red_blue_cross_fade = 0x2E
+    green_blue_cross_fade = 0x2F
+    seven_color_strobe_flash = 0x30
+    red_strobe_flash = 0x31
+    green_strobe_flash = 0x32
+    blue_strobe_flash = 0x33
+    yellow_strobe_flash = 0x34
+    cyan_strobe_flash = 0x35
+    purple_strobe_flash = 0x36
+    white_strobe_flash = 0x37
+    seven_color_jumping = 0x38
+
+    @staticmethod
+    def valid(pattern):
+        if pattern >= 0x25 and pattern <= 0x38 or pattern >= 0x61 and pattern <= 0x63:
+            return True
+        return False
+
+    @staticmethod
+    def valtostr(pattern):
+        for key, value in PresetPattern.__dict__.items():
+            if type(value) is int and value == pattern:
+                return key.replace("_", " ").title()
+        return None
